Change WAF configuration schema from singleton class to Config class
