+++ projects/packages/waf/src/class-waf-initializer.php
@@ -1,0 +1,39 @@
+<?php
+/**
+ * Class use to initialize the WAF module.
+ *
+ * @package automattic/jetpack-waf
+ */
+
+namespace Automattic\Jetpack\Waf;
+
+/**
+ * Initializes the module
+ */
+class Waf_Initializer {
+	/**
+	 * Initializes the configurations needed for the waf module.
+	 *
+	 * @return void
+	 */
+	public static function init() {
+		add_action( 'jetpack_activate_module_waf', __CLASS__ . '::on_activation' );
+		add_action( 'jetpack_deactivate_module_waf', __CLASS__ . '::on_deactivation' );
+	}
+
+	/**
+	 * On module activation set up waf mode
+	 */
+	public function on_activation() {
+		update_option( Waf_Runner::MODE_OPTION_NAME, 'normal' );
+		Waf_Runner::activate();
+		( new Waf_Standalone_Bootstrap() )->generate();
+	}
+
+		/**
+		 * On module deactivation, unset waf mode
+		 */
+	public function on_deactivation() {
+		Waf_Runner::deactivate();
+	}
+}
